+++ core/boot/src/main/java/org/eclipse/dataspaceconnector/boot/system/injection/ReflectiveObjectFactory.java
@@ -1,10 +1,23 @@
-package org.eclipse.dataspaceconnector.boot.system;
+/*
+ *  Copyright (c) 2022 Microsoft Corporation
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Microsoft Corporation - initial API and implementation
+ *
+ */
+package org.eclipse.dataspaceconnector.boot.system.injection;
 
 import org.eclipse.dataspaceconnector.spi.EdcException;
-import org.eclipse.dataspaceconnector.spi.system.InjectionContainer;
-import org.eclipse.dataspaceconnector.spi.system.InjectionPointScanner;
-import org.eclipse.dataspaceconnector.spi.system.Injector;
-import org.eclipse.dataspaceconnector.spi.system.ObjectFactory;
+import org.eclipse.dataspaceconnector.spi.system.injection.InjectionContainer;
+import org.eclipse.dataspaceconnector.spi.system.injection.InjectionPointScanner;
+import org.eclipse.dataspaceconnector.spi.system.injection.Injector;
+import org.eclipse.dataspaceconnector.spi.system.injection.ObjectFactory;
 import org.eclipse.dataspaceconnector.spi.system.ServiceExtensionContext;
 import org.jetbrains.annotations.NotNull;
