+++ spi/core-spi/src/main/java/org/eclipse/dataspaceconnector/spi/system/injection/InjectionPointScanner.java
@@ -1,4 +1,20 @@
-package org.eclipse.dataspaceconnector.spi.system;
+/*
+ *  Copyright (c) 2022 Microsoft Corporation
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Microsoft Corporation - initial API and implementation
+ *
+ */
+package org.eclipse.dataspaceconnector.spi.system.injection;
+
+import org.eclipse.dataspaceconnector.spi.system.Feature;
+import org.eclipse.dataspaceconnector.spi.system.Inject;
 
 import java.util.Arrays;
 import java.util.Set;
